backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary~=2.9.11
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==25.1.0
python-multipart==0.0.6

# HTTP client for Bria API
httpx==0.25.2
aiofiles==23.2.1

# Image processing
Pillow==10.1.0

# Configuration and environment
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
hypothesis==6.92.1

# Logging
python-json-logger==2.0.7

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1